fix: multiple data types without swagger keywords
The JSDoc syntax for defining multiple data types for a property was
already supported, but only when the user did explicitely add one of the
following keywords as the first item in the type list: 'oneOf', 'allOf'
or 'anyOf'.

Scenarios where the user did just use the default JSDoc notation without
using those keywords, like '@property {string|number}', were not
properly working, though.

In order to support this, we now check if the first item is one of the
keywords: if it's not, we'll assume that the first item contains also a
data type when building the schema. For those cases, 'oneOf' will be
used by default to wrap that type list.