Update path transform to make the security key optional (#179)
* Update path transform to make the security key optional

If the security key is present in a path swagger will take that definition over the global security one. By always adding this key we are disabling security globally. Modified the transform for a path to only set the key when values are found.

* fix(linting): error not related to my change ???? forced by commit hook
feat(security): Added the ability to disable security

By adding an empty @security tag to the jsdoc block for a path

* fix(test): test failed after making an unrelated change

These tests broke. I'm not sure why only making a change in
transforms/paths/index.js under the parsePath response
triggers the failures.

* fix(test): fixed tests and added a couple

Co-authored-by: Kevin Julián Martínez Escobar <kevinccbsg@gmail.com>